+++ lib/constants.js
@@ -8,6 +8,7 @@ export const RESERVED_PATHS = [
   "contact",
   "contact-us",
   "contribute",
+  "edit",
   "home",
   "login",
   "logout",
