+++ lib/transcripts.js
@@ -111,12 +111,15 @@ export async function getTranscriptData(id) {
   else {
     const contentHtml = await getMarkdownAsHTML(doc.data().transcript);
     return {
+      id: id,
       name: doc.data().name,
       contentHtml,
       link: doc.data().link,
       hashtags: doc.data().hashtags,
       creatorName: doc.data().creatorName || null,
       creatorLink: doc.data().creatorLink || null,
+      uid: doc.data().uid || null,
+      searchable: doc.data().searchable || false
     };
   }
 }
@@ -151,4 +154,12 @@ export async function createTranscript(url, data, successCB, failCB) {
   } else {
     if (failCB) failCB();
   }
+}
+
+export async function updateTranscript(url, data, successCB, failCB) {
+  // add to transcript collection
+  const docRef = await ref.collection("transcripts").doc(url);
+  return docRef.update(data)
+    .then(successCB)
+    .catch(failCB);
 }
